refactor(executable): integrated into remoteCE
